+++ src/TFLidar.h
@@ -7,8 +7,8 @@
 class TFLidar{
   public:
     TFLidar(TFBase *TF_Lidar);
-    void begin(SoftwareSerial *TFSerial , uint16_t baud_rate = 115200);
-    void begin(HardwareSerial *TFSerial , uint16_t baud_rate = 115200);
+    void begin(SoftwareSerial *TFSerial , unsigned long baud_rate = 115200);
+    void begin(HardwareSerial *TFSerial , unsigned long baud_rate = 115200);
     uint16_t get_distance(void);
     uint16_t get_strength(void);
     uint16_t get_chip_temperature(void);
